+++ packages/ui/src/themes/RelayKitTheme.ts
@@ -29,6 +29,9 @@ export interface RelayKitTheme {
       color?: string
       background?: string
     }
+    cta?: {
+      fontStyle?: string
+    }
   }
   input?: {
     background?: string
@@ -78,8 +81,8 @@ export interface RelayKitTheme {
 }
 
 export const defaultTheme: RelayKitTheme = {
-  font: 'Barlow, -apple-system, Helvetica, sans-serif',
-  fontHeading: 'Chivo, -apple-system, Helvetica, sans-serif',
+  font: 'Inter, -apple-system, Helvetica, sans-serif',
+  fontHeading: 'Inter, -apple-system, Helvetica, sans-serif',
   primaryColor: 'primary9',
   focusColor: 'primary7',
   subtleBackgroundColor: 'gray1',
@@ -110,6 +113,9 @@ export const defaultTheme: RelayKitTheme = {
     disabled: {
       color: 'gray11',
       background: 'gray8'
+    },
+    cta: {
+      fontStyle: 'normal'
     }
   },
   input: {